Flask Web Gateway/requirements.txt
flask
flask_cors
waitress
python-dotenv
psycopg2-binary
pandas